Ensure all output is printed when bootkube fails
It seems that remote-exec in Terraform does not wait for output buffer
to flush before exiting, so when execution fails, it sometimes happens,
that not entire output gets printed, which makes debugging more
difficult.

This issue has been reported upstraem:
https://github.com/hashicorp/terraform/issues/27121

As a workaround, this commit adds 'stdbuf -i0 -o0 -e0' to 'journalctl'
command, which disables standard output buffering, which ensures that
all output is flushed before executing next command, so returning an
error to the Terraform executor.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>